backend/requirements.txt
django==5.1.2
djangorestframework==3.15.2
pytest-django==4.9.0
psycopg2-binary==2.9.10
dataclasses-json==0.6.7
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.3.1
adrf==0.1.9
aiohttp==3.11.11
uvicorn==0.34.0
asyncio==3.4.3
pytest-asyncio==0.25.3